1. Категории. Переработка функциональности с учетом добавления атрибута id в сущность Catalog. При работе с памятью значение генерируется с помощью класса Random, при работе с БД - с помощью последовательности на уровне БД. 2. Репозиторий категорий. Реализована архитектура с возможностью настройки имплементации (свойство repository.mode): работа с памятью (repository.mode=memory) или работа с БД Postgres (repository.mode=postgres). Класс CategoryRepository преобразован в интерфейс. 3. Конфигурирование Hibernate и использование на уровне репозитория категорий. 4. Адаптация тестов по сервисам CatalogService и CategoryService.

+++ src/main/java/com/colvir/shop/repository/impl/CatalogRepositoryPostgresImpl.java
@@ -16,18 +16,18 @@ import java.util.Set;
 @Getter
 @RequiredArgsConstructor
 public class CatalogRepositoryPostgresImpl implements CatalogRepository {
-    // TODO. Пока это копия CatalogRepositoryMemoryImpl. Необходимо адаптировать на работу с БД
+    // Реализация с использованием JdbcTemplate
 
     private final JdbcTemplate jdbcTemplate;
 
     private final BeanPropertyRowMapper<Catalog> beanPropertyRowMapper = new BeanPropertyRowMapper<>(Catalog.class);
 
-    private Long generateId() {
+    private Integer generateId() {
 
         String statementString = "SELECT nextval('sequence_catalog_id')";
         SqlRowSet sqlRowSet = jdbcTemplate.queryForRowSet(statementString);
         sqlRowSet.next();
-        return sqlRowSet.getLong(1);
+        return sqlRowSet.getInt(1);
     }
 
     public Catalog save(Catalog catalog) {
